Insert some explicit casts to squash warnings.

+++ src/runtime/runtime.ptx_host.cpp
@@ -52,7 +52,7 @@ extern "C" {
     cuEventSynchronize(__end);                              \
     float msec;                                             \
     cuEventElapsedTime(&msec, __start, __end);              \
-    halide_printf("   (took %fms, t=%lld)\n", msec, halide_current_time_ns()); \
+    halide_printf("   (took %fms, t=%lld)\n", msec, (long long)halide_current_time_ns()); \
 } while(0)
 #endif //DEBUG
 
@@ -277,7 +277,7 @@ WEAK void halide_init_kernels(const char* ptx_src) {
         }
 
         #ifdef DEBUG
-        halide_printf("Got device %d, about to create context (t=%lld)\n", dev, halide_current_time_ns());
+        halide_printf("Got device %d, about to create context (t=%lld)\n", dev, (long long)halide_current_time_ns());
         #endif
 
 
@@ -352,7 +352,7 @@ static CUfunction __get_kernel(const char* entry_name)
 
     #ifdef DEBUG
     char msg[256];
-    snprintf(msg, 256, "get_kernel %s (t=%lld)", entry_name, halide_current_time_ns() );
+    snprintf(msg, 256, "get_kernel %s (t=%lld)", entry_name, (long long)halide_current_time_ns() );
     #endif
 
     // Get kernel function ptr
@@ -400,7 +400,8 @@ WEAK void halide_copy_to_dev(buffer_t* buf) {
         size_t size = buf_size(buf);
         #ifdef DEBUG
         char msg[256];
-        snprintf(msg, 256, "copy_to_dev (%zu bytes) %p -> %p (t=%lld)", size, buf->host, (void*)buf->dev, halide_current_time_ns() );
+        snprintf(msg, 256, "copy_to_dev (%zu bytes) %p -> %p (t=%lld)",
+                 size, buf->host, (void*)buf->dev, (long long)halide_current_time_ns() );
         halide_assert(halide_validate_dev_pointer(buf));
         #endif
         TIME_CALL( cuMemcpyHtoD(buf->dev, buf->host, size), msg );
@@ -445,7 +446,7 @@ WEAK void halide_dev_run(
         msg, 256,
         "dev_run %s with (%dx%dx%d) blks, (%dx%dx%d) threads, %d shmem (t=%lld)",
         entry_name, blocksX, blocksY, blocksZ, threadsX, threadsY, threadsZ, shared_mem_bytes,
-        halide_current_time_ns()
+        (long long)halide_current_time_ns()
     );
     #endif
